:sparkles: [Feat] 카드 리스트 관리 파일 별도 생성
index에서 card를 직접 호출하는 대신 allcardsection을 통해 card list를 호출하게 변경
css 파일 분리
