+++ src/components/challenge/AllCardSection.module.css
@@ -1,0 +1,47 @@
+.challengeTableWrapper {
+  width: 100%;
+  overflow-x: auto;
+  padding-bottom: 20px;
+  box-sizing: border-box;
+}
+
+.AllCardSection {
+  margin-top: 24px;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  gap: 24px;
+}
+
+.paginationWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 40px !important;
+  width: 100%;
+}
+
+.noChallenges {
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 19.09px;
+  text-align: center;
+  color: var(--grey-500);
+  margin-top: 233px;
+}
+
+/* 모바일 버전 */
+@media (max-width: 743px) {
+  .AllCardSection {
+    margin-top: 24px;
+    display: flex;
+    flex-direction: column;
+    box-sizing: border-box;
+    gap: 16px;
+  }
+
+  .noChallenges {
+    font-size: 14px;
+    margin-top: 223px;
+  }
+}
